Add something to the error queue when Proc-Type version is wrong
The Proc-Type header starts with a version number. From RFC 1421,
Section 4.6.1.1:

>  The "Proc-Type:" field has two subfields, separated by a comma.  The
>  first subfield is a decimal number which is used to distinguish among
>  incompatible encapsulated header field interpretations which may
>  arise as changes are made to this standard.  Messages processed
>  according to this RFC will carry the subfield value "4" to
>  distinguish them from messages processed in accordance with prior PEM
>  RFCs.

RFC 1421 was an update of RFC 1113, which used X-Proc-Type: 3,ENCRYPTED,
which we do not support. RFC 1040 defined X-Proc-Type: 2 and RFC 989
defined X-Proc-Type: 1,E. As far as I can tell, no other numbers for
Proc-Type have ever existed, and likely never will.

If we were to ever Proc-Type: 5, we currently return failure but forget
to put something on the error queue. It's possible this was originally
done for extensibility, but it just breaks internal invariants. By
returning zero, the callers will treat this as an error anyway. We'll
just confuse code that expects OpenSSL to return an error. Also
OpenSSL's PEM APIs all treat failure to decrypt (e.g. due to unsupported
future cipher) as an error, so we should treat an unsupported future PEM
encryption scheme as an error too.

Change-Id: Ia1f4f6776fea8d8a465a73105bf3ce24a587f26b
Reviewed-on: https://boringssl-review.googlesource.com/c/boringssl/+/74809
Commit-Queue: David Benjamin <davidben@google.com>
Reviewed-by: Adam Langley <agl@google.com>